Move PreparacaoPage to presentation and fix imports
