+++ test/widget_test.dart
@@ -10,7 +10,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:relic_ttprod/app.dart';
-import 'package:relic_ttprod/features/preparacao/presentation/widgets/preparacao_page.dart';
+import 'package:relic_ttprod/features/preparacao/presentation/preparacao_page.dart';
 
 void main() {
   testWidgets('App builds home page', (WidgetTester tester) async {
